requirements.txt
flask 
requests 
flask-wtf
enum
